Figured out Categories section and polished the rest of the project
