+++ .env
@@ -1,0 +1,7 @@
+# The .env file is used to store environment variables for the React application.
+# These variables are accessible in the application code using process.env.
+
+# REACT_APP_NEWS_API_KEY stores the API key for the news service being used.
+# By convention, all custom environment variables in a Create React App project
+# should start with REACT_APP_ to ensure they are properly embedded in the build.
+REACT_APP_NEWS_API_KEY = "ce478bc208b34d6f884c66f4be51c069"
