+++ src/NewsContext.js
@@ -1,16 +1,20 @@
 import { createContext, useState, useEffect } from 'react';
-import { fetchTopHeadlines } from './newsAPI';
+import { fetchTopHeadlines } from './utils/newsAPI';
 
 const NewsContext = createContext();
 
 export default NewsContext;
 
+// NewsProvider component serves as a wrapper to manage the state and logic
+// related to fetching and filtering news items.
 export const NewsProvider = ({ children }) => {
+  // State variables for news items, selected country, search filter and filtered news items
   const [newsItems, setNewsItems] = useState([]);
   const [selectedCountry, setSelectedCountry] = useState('gb'); // Default selected country is "gb"
   const [searchFilter, setSearchFilter] = useState('');
   const [filteredNewsItems, setFilteredNewsItems] = useState([]);
 
+  // useEffect to fetch news based on the selected country
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -32,6 +36,7 @@ export const NewsProvider = ({ children }) => {
     fetchData();
   }, [selectedCountry]);
 
+  // useEffect to filter news items based on the search filter
   useEffect(() => {
     const filterNewsItems = () => {
       const filteredItems = newsItems.filter(
@@ -42,10 +47,10 @@ export const NewsProvider = ({ children }) => {
       setFilteredNewsItems(filteredItems);
     };
 
-
     filterNewsItems();
   }, [newsItems, searchFilter]);
 
+  // Functions to update selected country, search filter, and news items
   const updateSelectedCountry = (country) => {
     setSelectedCountry(country);
   };
@@ -58,6 +63,7 @@ export const NewsProvider = ({ children }) => {
     setNewsItems(articles);
   };
 
+  // Context value containing state variables and update functions
   const contextValue = {
     newsItems,
     selectedCountry,
@@ -68,6 +74,7 @@ export const NewsProvider = ({ children }) => {
     updateNewsItems,
   };
 
+  // Providing the context value to children components
   return (
     <NewsContext.Provider value={contextValue}>
       {children}
