--- src/NewsDisplay.css
@@ -1,82 +1,0 @@
-.news-display {
-    display: flex;
-    flex-direction: column;
-}
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px;
-    background-color: lightgray;
-}
-
-.left-section {
-    display: flex;
-}
-
-.right-section {
-    display: flex;
-}
-
-.news-item {
-    border: 1px solid lightgray;
-    padding: 10px;
-    margin-top: 5px;
-    margin-bottom: 5px;
-}
-
-.search-form {
-    display: flex;
-    margin-bottom: 10px;
-    justify-content: center;
-}
-
-.search-form input {
-    margin-right: 10px;
-}
-
-.search-form button {
-    padding: 5px 10px;
-}
-
-.active {
-    background-color: lightblue;
-}
-
-.news-row {
-    display: flex;
-    justify-content: center;
-    margin: 5px;
-}
-
-.news-column {
-    display: flex;
-    flex-direction: column;
-    margin-left: 5px;
-    margin-right: 5px;
-}
-
-.news-item a {
-    text-align: right;
-    display: block;
-    margin-top: 10px;
-}
-
-.news-item-image {
-    width: 100%;
-    height: auto;
-    margin-bottom: 10px;
-}
-
-.news-item-more {
-    text-align: right;
-    display: block;
-    margin-top: 10px;
-}
-
-@media (max-width: 100%) {
-    .news-item {
-        width: 100%;
-    }
-}
