--- src/config.js
@@ -1,1 +1,0 @@
-export const API_KEY = 'd1336defe5894e02b0c21d5ebc628e0a';
