+++ src/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
+// Create a root using the ReactDOM.createRoot method
 const root = ReactDOM.createRoot(document.getElementById('root'));
+
+// Render the App component inside React.StrictMode using the root
 root.render(
   <React.StrictMode>
     <App />
