--- src/placeholderDataUS.js
@@ -1,40 +1,0 @@
-export const placeholderResponse = {
-  articles: [
-    {
-      title: "Sample Article 1",
-      image: "https://example.com/image1.jpg",
-      description: "This is a placeholder article description.",
-      url: "https://www.wmbfnews.com/sports/betting/2023/05/12/lakers-vs-warriors-western-conference-semifinals-game-6-player-prop-bets/",
-    },
-    {
-      title: "Sample Article 2",
-      image: "https://example.com/image1.jpg",
-      description: "This is another placeholder article description.",
-      url: "https://example.com/article2",
-    },
-    {
-      title: "Sample Article 1",
-      image: "",
-      description: "This is a placeholder article description.",
-      url: "https://www.wmbfnews.com/sports/betting/2023/05/12/lakers-vs-warriors-western-conference-semifinals-game-6-player-prop-bets/",
-    },
-    {
-      title: "Sample Article 2",
-      image: "https://example.com/image1.jpg",
-      description: "This is another placeholder article description.",
-      url: "https://example.com/article2",
-    },
-    {
-      title: "Sample Article 1",
-      image: "",
-      description: "This is a placeholder article description.",
-      url: "https://www.wmbfnews.com/sports/betting/2023/05/12/lakers-vs-warriors-western-conference-semifinals-game-6-player-prop-bets/",
-    },
-    {
-      title: "Sample Article 2",
-      image: "https://example.com/image1.jpg",
-      description: "This is another placeholder article description.",
-      url: "https://example.com/article2",
-    },
-  ],
-};
