+++ src/styles/index.css
@@ -1,0 +1,20 @@
+/* The body selector styles the entire body of the document */
+body {
+  margin: 0;
+  /* Removes default margin around the body */
+  /* Sets the font-family for the body to a list of fonts, falling back to the next font if the previous one is not available */
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  /* Enables font smoothing for better readability on Webkit browsers (e.g., Safari, Chrome) */
+  -webkit-font-smoothing: antialiased;
+  /* Enables font smoothing for better readability on Firefox */
+  -moz-osx-font-smoothing: grayscale;
+}
+
+/* The code selector styles the text within a <code> element */
+code {
+  /* Sets the font-family for the <code> element to a list of monospace fonts, falling back to the next font if the previous one is not available */
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
